--- docs/source/scratch.txt
@@ -1,102 +1,0 @@
-    if len(params) == 0:
-        return
-
-    grouped_tensors = _group_tensors_by_device_dtype_and_is_multidim(
-        [params, grads, row_vars, col_vars, variances, state_steps]  # type: ignore[list-item]
-    )
-    for (device, dtype, is_multidim), (
-        (
-            device_params,
-            device_grads,
-            device_row_vars,
-            device_col_vars,
-            device_variances,
-            device_state_steps,
-        )
-    ) in grouped_tensors.items():
-        if eps1 is None:
-            eps1 = torch.finfo(dtype).eps
-
-        if maximize:
-            device_grads: List[Tensor] = torch._foreach_neg(device_grads)
-
-        # Update steps
-        # If steps are on CPU, foreach will fall back to the slow path, which is a for-loop calling t.add(1) over
-        # and over. 1 will then be wrapped into a Tensor over and over again, which is slower than if we just
-        # wrapped it once now. The alpha is required to assure we go to the right overload.
-        if not torch._utils.is_compiling() and device_state_steps[0].is_cpu:
-            torch._foreach_add_(
-                device_state_steps, torch.tensor(1.0, device="cpu"), alpha=1.0
-            )
-        else:
-            torch._foreach_add_(device_state_steps, 1)
-
-        one_minus_beta2_ts = torch._foreach_pow(device_state_steps, beta2_decay)
-        one_minus_beta2_ts: List[Tensor] = [
-            v.to(device=device, dtype=dtype) for v in one_minus_beta2_ts
-        ]
-        rho_ts = [min(lr, 1 / (s.item() ** 0.5)) for s in device_state_steps]
-        alphas = [
-            max(eps2, p.norm(2).item() / (p.numel() ** 0.5)) * r
-            for p, r in zip(device_params, rho_ts)
-        ]
-
-        # Perform stepweight decay
-        if weight_decay != 0:
-            torch._foreach_mul_(device_params, 1 - lr * weight_decay)
-
-        if is_multidim:
-            device_row_vars: List[Tensor] = device_row_vars
-            device_col_vars: List[Tensor] = device_col_vars
-            assert (
-                device_row_vars[0] is not None and device_col_vars[0] is not None
-            ), "row_var and col_var should be defined when grad is multidimensional"
-            # same as (g * g).mean(dim=-1) w/o materializing an intermediate size g
-            row_means: List[Tensor] = [
-                torch.norm(grad, dim=-1, keepdim=True) for grad in device_grads
-            ]
-            torch._foreach_mul_(row_means, row_means)
-            torch._foreach_div_(row_means, [grad.size(-1) for grad in device_grads])
-            torch._foreach_lerp_(device_row_vars, row_means, one_minus_beta2_ts)
-
-            # same as (g * g).mean(dim=-2) w/o materializing an intermediate size g
-            col_means = [
-                torch.norm(grad, dim=-2, keepdim=True) for grad in device_grads
-            ]
-            torch._foreach_mul_(col_means, col_means)
-            torch._foreach_div_(col_means, [grad.size(-2) for grad in device_grads])
-            torch._foreach_lerp_(device_col_vars, col_means, one_minus_beta2_ts)
-
-            var_estimates: List[Tensor] = [
-                row_var @ col_var
-                for row_var, col_var in zip(device_row_vars, device_col_vars)
-            ]
-            row_var_means: List[Tensor] = [
-                row_var.mean(dim=-2, keepdim=True) for row_var in device_row_vars
-            ]
-            torch._foreach_clamp_min_(row_var_means, eps1)
-            torch._foreach_div_(var_estimates, row_var_means)
-        else:
-            device_variances: List[Tensor] = device_variances
-            assert (
-                device_variances[0] is not None
-            ), "variance should be defined when grad is a vector"
-
-            grads_squared: List[Tensor] = torch._foreach_mul(device_grads, device_grads)
-            torch._foreach_lerp_(device_variances, grads_squared, one_minus_beta2_ts)
-            # avoid writing into variance during update
-            var_estimates = [v.clone() for v in device_variances]
-
-        # square the eps1 as we sqrt after to keep eps1's magnitude
-        torch._foreach_clamp_min_(var_estimates, eps1 * eps1)
-        torch._foreach_sqrt_(var_estimates)
-        torch._foreach_reciprocal_(var_estimates)
-        torch._foreach_mul_(var_estimates, device_grads)
-        updates: List[Tensor] = var_estimates
-
-        alphas = [
-            -a / (max(1.0, update.norm(2).item() / ((update.numel() ** 0.5) * d)))
-            for a, update in zip(alphas, updates)
-        ]
-        torch._foreach_mul_(updates, alphas)
-        torch._foreach_sub_(device_params, updates)
