--- torch/optim/muon.py
@@ -1,4 +1,0 @@
-from torch.optim._muon import Muon, muon
-
-
-__all__ = ["Muon", "muon"]
